Fixed the Load operations incompatibility (and added more comments to Controller.sv)
For some reason, raddress was set to get the exact byte address (unlike
how waddress worked, and thus the correct implementation of the new
write instructions). Memoria32data is built to fetch the byte from word
alligned adresses, and, as it was, it was receiveing a word UNLIGNED
address from raddress in datamemory.sv (UNLIKE WADDRESS, FOR SOME
REASON!!!!!). (EXPLAINING THE BUG: let's say you chose the address 5.
datamemory would set the WORD-ALIGNED address to 5, Memoria32data would
return the bytes of address 8, 7, 6 and 5 instead of 7, 6, 5, and 4, and
then datamemory would get the offset of 1 from the incorrect
word-alligned address recieved, meaning, it would send the byte of
address 6 to the register).