JAL and JALR implementation
Adds support for JAl and JALR functions, enabling unconditional jumps
and subroutine calls.
KEY CHANGES:
1. JAL implementations:
- Jump logic: A new Jump control signal is generated by the Controller
to trigger the PC update in the BranchUnit, reusing the existing branch
hardware (adapting the adder in the BranchUnit module) for unconditional
jumps.
- Link logic: MemtoReg control was widened to 2-bits to be able to
support the write-back of the PC+4 address (It only was able to choose
the register write data from the ALU or the data memory with one bit).
With this change, the resmux had to be replaced with a 4-input mux (for
the new third write-back option).

2. JALR implementations:
- Jump logic: The ALU is used to calculate the register1 + imm target
address. A new JumpR control signal is generated by the Controller to
trigger the PC update in the BranchUnit, also reusing the adder module,
but utilizing the result (with the least significant bit cleared, as per
the ISA specification) from the ALU (ALUResult) as the jump target,
instead of the already calculated PC+IMM.
- Link logic: Reused logic from the JAL operation.

SUMMARY OF MODULE CHANGES:
Controller.sv: Decodes the new JAL and JALR opcodes and generates the
new Jump and Jumpr control signals. Manages new extended MemtoReg
control logic, updating the RegWrite, ALUSrc, and ALUOp logic.
imm_Gen.sv: Extended to correctly concatenate the J-type instruction
immediate (for JAL) and the (new) I-type immediate (for JALR).
BranchUnit.sv: Logic updated to trigger PC update with Jump and JumpR
and select the correct target address(PC + IMM for JAL, PC + ALUResult
for JALR).
RegPack.sv: Register definitions updated to carry new Jump and JumpR
signals and the extended MemtoReg signal.
Datapath.sv: Final mux for writeback replaced (to support new 2-bit
selector RegtoMem), and new control signals correctly wired through the
pipeline stages.